// frontend/src/app/shared/generated-api/model/userProfile.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserProfile { 
    id: string;
    username: string;
    email: string;
    displayName?: string;
    role: UserProfileRoleEnum;
    emailChangeAvailable: boolean;
    selfDeletionEnabled: boolean;
}
export enum UserProfileRoleEnum {
    User = 'USER',
    Admin = 'ADMIN'
};